+++ system/src/Grav/Common/Service/PagesServiceProvider.php
@@ -80,6 +80,8 @@ class PagesServiceProvider implements ServiceProviderInterface
                 }
                 // Default route test and redirect
                 if ($config->get('system.pages.redirect_default_route') && $page->route() !== $path) {
+                    $uri->setUriProperties(['path' => $page->route()]);
+                    $url = (string) $uri;
                     $c->redirect($url);
                 }
             }