+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 1. [](#bugfix)
     * Rework logic to pull out excluded files from pipeline more reliably [#2445](https://github.com/getgrav/grav/issues/2445)
     * Better logic in `Utils::normalizePath` to handle externals properly [#2216](https://github.com/getgrav/grav/issues/2216)
+    * Fix to force all `Page::taxonomy` to be treated as strings [#2446](https://github.com/getgrav/grav/issues/2446)
 
 # v1.6.3
 ## 04/12/2019